+++ frontend/src/Store/Actions/interactiveImportActions.js
@@ -216,7 +216,7 @@ export const actionHandlers = handleThunks({
         path: item.path,
         authorId: item.author ? item.author.id : undefined,
         bookId: item.book ? item.book.id : undefined,
-        foreignEditionId: item.foreignEditionId ? item.ForeignEditionId : undefined,
+        foreignEditionId: item.foreignEditionId || undefined,
         quality: item.quality,
         releaseGroup: item.releaseGroup,
         indexerFlags: item.indexerFlags,