+++ src/NzbDrone.Core/MediaFiles/InvalidBookFileLink.cs
@@ -1,0 +1,11 @@
+namespace NzbDrone.Core.MediaFiles
+{
+    public class InvalidBookFileLink
+    {
+        public int BookFileId { get; set; }
+        public string Path { get; set; }
+        public int EditionId { get; set; }
+        public int? BookId { get; set; }
+        public string Reason { get; set; }
+    }
+}